refactor: MAT-362 MatchDetailCard -> MatchDetailPanel로 컴포넌트 이름 일관성 있게 변경
