--- src/routes/teams/$teamId/matches/-components/MatchDetailCard/index.ts
@@ -1,1 +1,0 @@
-export * from './MatchDetailCard';
