+++ src/routes/teams/$teamId/matches/-components/MatchDetailPanel/GoalList/GoalList.css.ts
