+++ src/routes/teams/$teamId/matches/-components/MatchDetailPanel/GoalList/index.ts
