+++ src/routes/teams/$teamId/matches/-components/MatchDetailPanel/MatchDetailLink/MatchDetailLink.css.ts
