+++ src/routes/teams/$teamId/matches/-components/MatchDetailPanel/MatchDetailLink/index.ts
