+++ src/routes/teams/$teamId/matches/-components/MatchDetailPanel/MatchDetailPanel.css.ts
