+++ src/routes/teams/$teamId/matches/-components/MatchDetailPanel/index.ts
@@ -1,0 +1,1 @@
+export * from './MatchDetailPanel';
