+++ src/routes/teams/$teamId/matches/-components/index.ts
@@ -1,4 +1,4 @@
 export * from './TeamStatsSummary';
 export * from './MatchFilters';
 export * from './MatchTable';
-export * from './MatchDetailCard';
+export * from './MatchDetailPanel';
